LayoutTests/fast/dom/nonce-attribute-reflection-expected.txt

Tests that the nonce attribute for HTML link, script, and style elements is reflected.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS element = document.createElement("link"); element.setAttribute("nonce", "dummy"); element.nonce is "dummy"
PASS element = document.createElement("script"); element.setAttribute("nonce", "dummy"); element.nonce is "dummy"
PASS element = document.createElement("style"); element.setAttribute("nonce", "dummy"); element.nonce is "dummy"
PASS successfullyParsed is true

TEST COMPLETE
